refactor: reorganiza pacotes de DTOs para melhor estruturação
