+++ backend/src/main/java/br/com/gestorfinanceiro/dto/grafico/GraficoBarraDTO.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.dto;
+package br.com.gestorfinanceiro.dto.grafico;
 
 import java.math.BigDecimal;
 import java.util.Map;
