+++ backend/src/main/java/br/com/gestorfinanceiro/dto/user/LoginDTO.java
@@ -1,3 +1,3 @@
-package br.com.gestorfinanceiro.dto;
+package br.com.gestorfinanceiro.dto.user;
 
 public record LoginDTO(String email, String password) {}
