+++ backend/src/main/java/br/com/gestorfinanceiro/dto/user/UserDTO.java
@@ -1,7 +1,10 @@
-package br.com.gestorfinanceiro.dto;
+package br.com.gestorfinanceiro.dto.user;
 
 
-import jakarta.validation.constraints.*;
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Pattern;
+import jakarta.validation.constraints.Size;
 
 public class UserDTO {
