+++ backend/src/main/java/br/com/gestorfinanceiro/dto/user/UserForAdminDTO.java
@@ -1,4 +1,4 @@
-package br.com.gestorfinanceiro.dto;
+package br.com.gestorfinanceiro.dto.user;
 
 public class UserForAdminDTO {
     private String uuid;
