+++ backend/src/main/java/br/com/gestorfinanceiro/services/DespesaService.java
@@ -1,7 +1,7 @@
 package br.com.gestorfinanceiro.services;
 
-import br.com.gestorfinanceiro.dto.GraficoBarraDTO;
-import br.com.gestorfinanceiro.dto.GraficoPizzaDTO;
+import br.com.gestorfinanceiro.dto.grafico.GraficoBarraDTO;
+import br.com.gestorfinanceiro.dto.grafico.GraficoPizzaDTO;
 import br.com.gestorfinanceiro.models.DespesaEntity;
 
 import java.math.BigDecimal;
