+++ backend/src/test/java/br/com/gestorfinanceiro/controllers/AdminControllerTest/AdminControllerUnitTest.java
@@ -3,7 +3,7 @@ package br.com.gestorfinanceiro.controllers.AdminControllerTest;
 import br.com.gestorfinanceiro.TestDataUtil;
 import br.com.gestorfinanceiro.config.security.JwtFilter;
 import br.com.gestorfinanceiro.controller.AdminController;
-import br.com.gestorfinanceiro.dto.UserForAdminDTO;
+import br.com.gestorfinanceiro.dto.user.UserForAdminDTO;
 import br.com.gestorfinanceiro.mappers.Mapper;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.services.AdminService;
