+++ backend/src/test/java/br/com/gestorfinanceiro/controllers/AuthControllerTest/AuthControllerIntegrationTest.java
@@ -2,8 +2,8 @@ package br.com.gestorfinanceiro.controllers.AuthControllerTest;
 
 import br.com.gestorfinanceiro.TestDataUtil;
 import br.com.gestorfinanceiro.controller.AuthController;
-import br.com.gestorfinanceiro.dto.LoginDTO;
-import br.com.gestorfinanceiro.dto.UserDTO;
+import br.com.gestorfinanceiro.dto.user.LoginDTO;
+import br.com.gestorfinanceiro.dto.user.UserDTO;
 import br.com.gestorfinanceiro.exceptions.auth.login.InvalidPasswordException;
 import br.com.gestorfinanceiro.models.UserEntity;
 import br.com.gestorfinanceiro.repositories.UserRepository;
